Create vaccinated.sql
Functional Dependency:
{vaccinated_date, medicare_num -> dose_num, f_name, f_address, v_type}
In our ‘reasonable assumption’, we assumed that a person cannot get vaccinated
more than once per day. Therefore, if the ‘medicare_num’, which is Medicare Card number,
and the ‘vaccinated_date’ are fixed, the rest of attributes will be set. The only FD is about
those two which control the rest, therefore this is in BCNF.