Create query_2.sql
Get details of all the doctors who have been infected by COVID-19 in the past two weeks. Details include doctor’s first-name, last-name, date of infection, and the name of the facility that the doctor is currently working for. Results should be displayed sorted in ascending order by the facility name, then by the first-name of the doctor.